feat: refactor persistence layer classes into their own folder
